tests: added initialize tests

+++ packages/vesting/test/utils/Events.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.20;
 
 interface Events {
-    event Initialized(uint8 version);
+    event Initialized(uint64 version);
 }
